setup the qrcode v3 finalized
