+++ app/checks.py
@@ -7,7 +7,14 @@ from app.mongo_db_connection import (
     get_process_object,
 )
 
-from .helpers import register_user_access
+def register_user_access(process_steps, authorized_role, user):
+        """Once someone has made changes to their docs"""
+        for step in process_steps:
+            if step["stepRole"] == authorized_role:
+                for clone_map in step["stepDocumentCloneMap"]:
+                    if user in clone_map:
+                        clone_map["accessed"] = True
+                        break
 
 
 def is_finalized(item_id, item_type):
